# packages/openapi/api-spec.yaml
openapi: 3.0.3
info:
  title: FuelEU Control Tower API
  version: 1.0.0
  description: OpenAPI Specification for FuelEU Maritime Compliance Platform ensuring deterministic generation for the UI Next.js client.
servers:
  - url: http://localhost:8082
    description: Local development server
paths:
...
  /api/v1/marketplace/cb-rate:
    get:
      summary: Fetch latest indicative CB market rate
      operationId: getCbMarketplaceRate
      parameters:
        - in: query
          name: year
          required: false
          schema:
            type: integer
            format: int32
            default: 2025
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CbMarketplaceQuote'
  /api/v1/marketplace/purchase:
    post:
      summary: Record a CB marketplace purchase for a vessel-year
      operationId: purchaseCbMarketplace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CbMarketplacePurchaseRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CbMarketplacePurchaseResponse'
        '400':
          description: Bad Request
...
components:
  schemas:
...
    CbMarketplaceQuote:
      type: object
      properties:
        year:
          type: integer
          format: int32
        source:
          type: string
        symbol:
          type: string
        priceUsdPerToken:
          type: number
          format: double
          description: Indicative token spot in USD per tCO2eq token.
        fxEurUsd:
          type: number
          format: double
        priceEurPerGco2eq:
          type: number
          format: double
          description: Price in EUR per gCO2eq of compliance balance.
        priceEurPerCb:
          type: number
          format: double
          description: Price in EUR per CB (1 CB = 1 tCO2eq in marketplace convention).
        priceEurPerTco2eq:
          type: number
          format: double
        cbToGco2eq:
          type: number
          format: double
          description: Conversion constant for marketplace CB to compliance balance units (1 CB = 1,000,000 gCO2eq).
        disclaimer:
          type: string
    CbMarketplacePurchaseRequest:
      type: object
      required: [vesselId]
      properties:
        vesselId:
          type: string
          format: uuid
        amountCb:
          type: number
          format: double
          description: CB amount to buy (1 CB = 1 tCO2eq).
        amountGco2eq:
          type: number
          format: double
          description: Legacy field - amount in gCO2eq.
        year:
          type: integer
          format: int32
          default: 2025
    CbMarketplacePurchaseResponse:
      type: object
      properties:
        status:
          type: string
        message:
          type: string
        amountGco2eq:
          type: number
          format: double
          description: Purchased amount converted to compliance balance units (gCO2eq).
        amountCb:
          type: number
          format: double
          description: Purchased marketplace amount in CB/tCO2eq units.
        cbToGco2eq:
          type: number
          format: double
          description: Conversion constant for marketplace CB to compliance units.
        unitPriceEurPerCb:
          type: number
          format: double
        unitPriceEurPerGco2eq:
          type: number
          format: double
        totalCostEur:
          type: number
          format: double
        source:
          type: string
        symbol:
          type: string
...